Modifications for JDK21 compatibility
Minor code modifications for the following errors:

possible 'this' escape before subclass is fully initialized
non-transient instance field of a serializable class declared with a non-serializable type
implicit cast from long to int in compound assignment is possibly lossy

+++ core/src/main/java/org/apache/accumulo/core/fate/zookeeper/ZooSession.java
@@ -108,7 +108,7 @@ public class ZooSession {
    */
   static ZooKeeper connect(String host, int timeout, String scheme, byte[] auth, Watcher watcher) {
     final int TIME_BETWEEN_CONNECT_CHECKS_MS = 100;
-    int connectTimeWait = Math.min(10_000, timeout);
+    long connectTimeWait = Math.min(10_000, timeout);
     boolean tryAgain = true;
     long sleepTime = 100;
     ZooKeeper zooKeeper = null;
